pyproject.toml
[tool.poetry]
name = "print-backend"
version = "0.1.0"
description = ""
authors = ["Viktor Zagoruyko <v.a.zagoruyko.2@gmail.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
Django = "^5.1"
psycopg2-binary = "^2.9.11"
django-environ = "^0.12.0"
django-simple-history = "^3.10.1"
faker = "^38.0.0"
djangorestframework = "^3.16.1"
django-viewflow = "^2.2.13"
python-json-logger = "^4.0.0"
celery = "^5.5.3"
redis = "^7.0.1"
django-celery-beat = "^2.8.1"
pytest = "^9.0.1"
pytest-django = "^4.11.1"
factory-boy = "^3.3.3"
django-cors-headers = "^4.9.0"
gunicorn = "^23.0.0"
rollbar = "^1.3.0"
requests = "^2.32.5"
reportlab = "^4.4.4"
pillow = "^12.0.0"
python-barcode = "^0.16.1"
django-json-widget = "^2.1.0"
pdf2image = "^1.17.0"
svglib = "^1.6.0"

[tool.poetry.group.dev.dependencies]
black = "^24.0"
isort = "^5.13"
flake8 = "^7.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
